requirements.txt
paho-mqtt
python-dotenv
Adafruit_DHT